+++ src/makeLossPlots.py
@@ -151,6 +151,7 @@ def plotLosses(lossTypes: list[list[str]], history: dict, startFrom: int) -> Fig
     fig, axs = plt.subplots(nrows=numRowsCols, ncols=numRowsCols, sharex=True, figsize=(15, 15))
     epochs = range(len(history[lossTypes[0][0]]))
     for i, lt in enumerate(lossTypes):
+        logUtils.info("Plotting loss type " + str(lt))
         allDats = []
         ax = axs[i // numRowsCols][i % numRowsCols]
         for loss in lt: