+++ src/shapToBigwig.py
@@ -72,13 +72,12 @@ def writeBigWig(inH5: h5py.File, outFname: str) -> None:  # pylint: disable=too-
     logUtils.debug("Bigwig header" + str((sorted(bwHeader))))
     numRegions = inH5["coords_chrom"].shape[0]
     if isinstance(inH5["coords_chrom"][0], bytes):
-        logUtils.warning("You are using an old-style hdf5 file for importance scores. "
-            "Support for these files will be removed in BPReveal 5.0. "
+        logUtils.error("You are using an old-style hdf5 file for importance scores. "
+            "This has been deprecated since 4.0.0 and is an error as of 5.0.0 "
             "Instructions for updating: Re-calculate importance scores.")
-        coordsChrom = np.array(inH5["coords_chrom"].asstr())
-    else:
-        coordsChromIdxes = np.array(inH5["coords_chrom"])
-        coordsChrom = np.array([chromIdxToName[x] for x in coordsChromIdxes])
+        raise ValueError("Invalid importance score file format.")
+    coordsChromIdxes = np.array(inH5["coords_chrom"])
+    coordsChrom = np.array([chromIdxToName[x] for x in coordsChromIdxes])
     # Sort the regions.
     coordsStart = np.array(inH5["coords_start"])
     coordsEnd   = np.array(inH5["coords_end"])  # noqa